# Phase_II/specs/1-fastapi-backend/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  description: Multi-user todo application API with JWT authentication
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server (Railway)
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/signup:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Create a new user account with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User created successfully
                  user_id:
                    type: string
                    format: uuid
                    example: 550e8400-e29b-41d4-a716-446655440000
        '400':
          description: Validation error or duplicate email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidEmail:
                  value:
                    detail: Invalid email format
                weakPassword:
                  value:
                    detail: Password must be at least 8 characters
                duplicateEmail:
                  value:
                    detail: Email already registered
  /api/auth/signin:
    post:
      tags:
        - Authentication
      summary: User authentication
      description: Sign in with email and password to receive JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: Invalid email or password
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/signin
  schemas:
    UserCreate:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: John Doe
        email:
          type: string
          format: email
          maxLength: 255
          example: john.doe@example.com
        password:
          type: string
          minLength: 8
          format: password
          description: Must contain uppercase, lowercase, number, and special character
          example: SecurePass123!
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          example: SecurePass123!
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: john.doe@example.com
        name:
          type: string
          example: John Doe
        created_at:
          type: string
          format: date-time
          example: 2026-01-08T12:00:00Z
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT token for authentication
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          example: bearer
        user:
          $ref: '#/components/schemas/UserResponse'
...
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: Invalid email format
...